requirements.txt
python_pptx==0.6.21
requests==2.27.1